ingestion/requirements.txt
kaggle==1.6.17
pandas==2.2.2
google-cloud-storage==2.16.0
pyarrow==16.0.0